+++ pages/main/article/index.vue
@@ -2,7 +2,7 @@
 <div class="padding-l">
         <el-row class="padding-l">
             <el-col :span="12">
-                <span class="font-weight-blder"><h1>项目列表</h1></span>
+                <span class="font-weight-blder"><h1>文章管理</h1></span>
             </el-col>
         </el-row>
         <el-form :inline="true" v-model="search" class="background-color-minor margin-bottom-m padding-m">
@@ -20,34 +20,39 @@
                   @sort-change="handleSortChange"
                   :default-sort="{prop: 'name', order: 'descending'}"
                   class="col-12">
-            <el-table-column type="expand">
-                <template slot-scope="props">
-                    <div v-html="props.row.projectDescription"></div>
-                </template>
-            </el-table-column>
             <el-table-column prop="id"
-                             label="项目ID"
+                             label="文章ID"
                              sortable
                              width="120">
             </el-table-column>
-            <el-table-column prop="projectName"
-                             label="项目名称"
+            <el-table-column prop="title"
+                             label="文章标题"
                              sortable
-                             >
+                             width="120">
             </el-table-column>
-            <el-table-column prop="creationTime"
-                             label="创建时间"
+            <el-table-column prop="category.name"
+                             label="所属栏目"
+                             sortable
+                             width="120">
+            </el-table-column>
+            <el-table-column prop="author"
+                             width="200"
+                             label="作者">
+            </el-table-column>
+            <el-table-column prop="creationTimeStr"
+                             label="发布时间"
                              width="200">
-                     <template slot-scope="scope">
+                    <template slot-scope="scope">
                         <el-icon name="time"></el-icon>
-                        <span>{{ scope.row.creationTime|time}}</span>
+                        <span>{{ scope.row.creationTimeStr}}</span>
                     </template>
             </el-table-column>
             <el-table-column label="操作"
                              fixed="right"
                              width="400">
-                <template slot-scope="scope">
+                <template slot-scope="scope">                    
                     <el-button size="small" icon="el-icon-edit" @click="onEdit(scope.row)">编辑</el-button>
+                    <el-button size="small" icon="el-icon-sort-up" @click="onTop(scope.row)">{{scope.row.isTop?"默认":"置顶"}}</el-button>
                     <el-button size="small" type="danger" icon="el-icon-delete"  @click="onDelete(scope.row)">删除</el-button>
                 </template>
             </el-table-column>
@@ -112,37 +117,50 @@ export default{
             this.list.currentPage = val;
             this.loadData();
         },
+        onTop(model) {
+            var me = this;
+             axios.put(apiConfig.article_top,{
+                "id": model.id,
+                "isTop": !model.isTop
+            }).then(response =>{
+                model.isTop=!model.isTop;
+            });
+        },
         onAdd() {
             var me = this;
-            me.$loaderwindow("/main/project/edit?id=0", "创建项目")
+            me.$loaderwindow("/main/article/edit?id=0", "创建文章")
                 .then( model => {
-                    me.$message({ type: "success", message: "创建项目成功！" });
+                    me.$message({ type: "success", message: "创建文章成功！" });
                     me.loadData();
                 });
         },
         onEdit(model) {
             var me = this;
-            me.$loaderwindow(`/main/project/edit?id=${model.id}`, `编辑项目${model.projectName}`)
-                .then(m => {
-                    me.$message({ type: "success", message: `编辑项目${model.projectName}成功！`});
+            me.$loaderwindow(`/main/article/edit?id=${model.id}`, "编辑文章")
+                .then(model => {
+                    me.$message({ type: "success", message: "编辑文章成功！" });
                     me.loadData();
                 });
         },
         onDelete(model) {
             var me = this;
-            me.$confirm(`是否永久删除【'${model.projectName}】?`, '询问', {
-                confirmButtonText: '删除',
-                cancelButtonText: '取消',
-                type: 'warning'
-            }).then(() => {
-                me.deleteSelect(model);
-            }).catch(() => {
+            if (model) {
+                me.$confirm('是否永久删除[' + model.userName + ']?', '询问', {
+                    confirmButtonText: '删除',
+                    cancelButtonText: '取消',
+                    type: 'warning'
+                }).then(() => {
+                    me.deleteSelect(model);
+                }).catch(() => {
 
-            });
+                });
+            } else {
+                me.$alert("请勾中要删除的项");
+            }
         },
         deleteSelect(model){
             var me=this;
-            axios.post(apiConfig.project_delete,{id:model.id}).then(response=>{
+            axios.delete(apiConfig.article_delete,{params:{id:model.id}}).then(response=>{
                 me.$message({ type: "success", message: "删除用户成功！" });
                 me.loadData();
             });
@@ -150,9 +168,9 @@ export default{
         loadData () {
             var me=this;
             me.list.loading = true;
-            axios.get(apiConfig.project_get, {params:{
+            axios.get(apiConfig.article_get, {params:{
                 skipCount: me.getSkip(),
-                maxCount: me.list.pageSize
+                maxResultCount: me.list.pageSize
             }})
             .then(response => {
                 me.list.tableData = response.data.result.items;
@@ -168,5 +186,4 @@ export default{
         me.loadData();
     }
 };
-</script>
-    
+</script>
